+++ deeplite_torch_zoo/src/dnn_blocks/yolo_blocks.py
@@ -7,6 +7,14 @@ from deeplite_torch_zoo.src.dnn_blocks.common import (ConvBnAct, DWConv,
                                                       round_channels)
 from deeplite_torch_zoo.src.dnn_blocks.resnet_blocks import (GhostBottleneck,
                                                              ResNeXtBottleneck)
+from torch.nn.modules.activation import LeakyReLU
+try:
+    from mish_cuda import MishCuda as Mish
+except:
+
+    class Mish(nn.Module):  # https://github.com/digantamisra98/Mish
+        def forward(self, x):
+            return x * torch.nn.functional.softplus(x).tanh()
 
 
 class SPPBottleneck(nn.Module):
@@ -87,6 +95,50 @@ class YOLOBottleneckCSP(nn.Module):
         y1 = self.cv3(self.m(self.cv1(x)))
         y2 = self.cv2(x)
         return self.cv4(self.act(self.bn(torch.cat((y1, y2), 1))))
+
+class YOLOBottleneckCSP2(nn.Module):
+    # CSP Bottleneck https://github.com/WongKinYiu/CrossStagePartialNetworks
+    def __init__(
+        self, c1, c2, n=1, shortcut=False, g=1, e=0.5, act='relu'
+    ):  # ch_in, ch_out, number, shortcut, groups, expansion
+        super().__init__()
+        c_ = int(c2)  # hidden channels
+        self.cv1 = ConvBnAct(c1, c_, 1, 1, act=act)
+        self.cv2 = nn.Conv2d(c_, c_, 1, 1, bias=False)
+        self.cv3 = ConvBnAct(2 * c_, c2, 1, 1, act=act)
+        self.bn = nn.BatchNorm2d(2 * c_)
+        self.act = Mish()
+        self.m = nn.Sequential(
+            *(YOLOBottleneck(c_, c_, shortcut=shortcut, g=g, e=1.0, act=act) for _ in range(n))
+        )
+
+    def forward(self, x):
+        x1 = self.cv1(x)
+        y1 = self.m(x1)
+        y2 = self.cv2(x1)
+        return self.cv3(self.act(self.bn(torch.cat((y1, y2), dim=1))))
+
+class YOLOBottleneckCSP2Leaky(nn.Module):
+    # CSP Bottleneck https://github.com/WongKinYiu/CrossStagePartialNetworks
+    def __init__(
+        self, c1, c2, n=1, shortcut=False, g=1, e=0.5, act='hswish',
+    ):  # ch_in, ch_out, number, shortcut, groups, expansion
+        super().__init__()
+        c_ = int(c2)  # hidden channels
+        self.cv1 = ConvBnAct(c1, c_, 1, 1, act=act)
+        self.cv2 = nn.Conv2d(c_, c_, 1, 1, bias=False)
+        self.cv3 = ConvBnAct(2 * c_, c2, 1, 1, act=act)
+        self.bn = nn.BatchNorm2d(2 * c_)
+        self.act = LeakyReLU(negative_slope=0.1)
+        self.m = nn.Sequential(
+            *[YOLOBottleneck(c_, c_, shortcut, g, e=1.0, act=act) for _ in range(n)]
+        )
+
+    def forward(self, x):
+        x1 = self.cv1(x)
+        y1 = self.m(x1)
+        y2 = self.cv2(x1)
+        return self.cv3(self.act(self.bn(torch.cat((y1, y2), dim=1))))
 
 
 class YOLOC3(nn.Module):
